+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -268,16 +268,27 @@ interface Synchronizer {
      */
     suspend fun getTransparentBalance(tAddr: String): WalletBalance
 
-    suspend fun getNearestRewindHeight(height: BlockHeight): BlockHeight
-
     /**
      * Returns the safest height to which we can rewind, given a desire to rewind to the height
      * provided. Due to how witness incrementing works, a wallet cannot simply rewind to any
      * arbitrary height. This handles all that complexity yet remains flexible in the future as
      * improvements are made.
      */
-    suspend fun rewindToNearestHeight(height: BlockHeight, alsoClearBlockCache: Boolean = false)
+    suspend fun getNearestRewindHeight(height: BlockHeight): BlockHeight
 
+    /**
+     * Rewinds to the safest height to which we can rewind, given a desire to rewind to the height
+     * provided. Due to how witness incrementing works, a wallet cannot simply rewind to any
+     * arbitrary height. This handles all that complexity yet remains flexible in the future as
+     * improvements are made.
+     */
+    suspend fun rewindToNearestHeight(height: BlockHeight)
+
+    /**
+     * Rewinds to the safest height approximately 14 days backward from the current chain tip. Due to how witness
+     * incrementing works, a wallet cannot simply rewind to any arbitrary height. This handles all that complexity
+     * yet remains flexible in the future as improvements are made.
+     */
     suspend fun quickRewind()
 
     /**