+++ CHANGELOG.md
@@ -1,14 +1,24 @@
 # Change Log
 
 ## Unreleased
-- `CompactBlockProcessor` now processes compact blocks from the lightwalletd server in a non-linear order. This 
-  feature shortens the time after which a wallet's spendable balance can be used.
+- `CompactBlockProcessor` now processes compact blocks from the lightwalletd server with spend-before-sync algorithm 
+  (i.e. non-linear order). This feature shortens the time after which a wallet's spendable balance can be used.
 
 ### Added
 - New `syncAlgorithm` parameter of `Synchronizer.new()` and `WalletCoordinator()` to select preferred 
   `CompactBlockProcessor` block synchronizing algorithm. It can be of `CompactBlockProcessor.SyncAlgorithm.LINEAR` 
-  or `NON_LINEAR`. The LINEAR type is automatically used if the client app does not specify otherwise. Please note 
-  that the NON_LINEAR type is currently unstable and still under development.
+  or `SPEND_BEFORE_SYNC`. The LINEAR type is automatically used if the client app does not specify otherwise. Please 
+  note that the SPEND_BEFORE_SYNC type is currently unstable and still under development.
+
+### Removed
+- `CompactBlockProcessor.ProcessorInfo.lastSyncHeight` which the SDK is no longer able to provide because of the new
+  **SpendBeforeSync** synchronization algorithm adoption. Use `CompactBlockProcessor.ProcessorInfo.overallSyncRange` 
+  which contains all blocks in case of `SpendBeforeSync` synchronization algorithm. No internal change was made in 
+  case of older Linear synchronization algorithm.
+- `CompactBlockProcessor.ProcessorInfo.isSyncing`. Use `Synchronizer.status` instead.
+- `CompactBlockProcessor.ProcessorInfo.syncProgress`. Use `Synchronizer.progress` instead.
+- `alsoClearBlockCache` parameter from rewind functions of `Synchronizer` and `CompactBlockProcessor` as it take no 
+  affect on the current rewind functionality result.
 
 ## 1.20.0-beta01
 - The SDK internally migrated from `BackendExt` rust backend extension functions to more type-safe `TypesafeBackend`.